# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/": {
            "get": {
                "description": "Welcome will return the welcome message as response with status code as 200",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System status"
                ],
                "summary": "Welcome",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Base"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "description": "Ping test, return the commit ID and container ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System status"
                ],
                "summary": "Ping test",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Base"
                        }
                    }
                }
            }
        },
        "/secret": {
            "post": {
                "description": "Add a new secret",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "secret"
                ],
                "summary": "Add a new secret",
                "parameters": [
                    {
                        "type": "string",
                        "description": "This text will be saved as a secret",
                        "name": "secret",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The secret won't be available after the given number of views. It must be greater than 0.",
                        "name": "expireAfterViews",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The secret won't be available after the given time. The value is provided in minutes. 0 means never expires",
                        "name": "expireAfter",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.SecretResponse"
                        }
                    }
                }
            }
        },
        "/secret/{hash}": {
            "get": {
                "description": "Get a secret by ID",
                "produces": [
                    "application/json",
                    "text/xml"
                ],
                "tags": [
                    "secret"
                ],
                "summary": "Get a secret by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique hash to identify the secret",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.SecretResponse"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Status will provide the status of the service we been used in the application",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System status"
                ],
                "summary": "Get Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Base"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Base": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.SecretResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "remainingViews": {
                    "type": "integer"
                },
                "secretText": {
                    "type": "string"
                }
            }
        }
    }
}